backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "liims-backend"
version = "0.1.0"
description = "LIIMS - Longevity India Information Management System API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "bcrypt>=4.1.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "weasyprint>=61.0",
    "python-docx>=1.1.0",
    "qrcode[pil]>=7.4.0",
    "httpx>=0.26.0",
    "python-dateutil>=2.8.0",
    "psycopg[binary]>=3.1.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
